Fix splitting of training data and evaluation

+++ commands/model_merge.py
@@ -3,7 +3,7 @@ from configs import *
 from commands.model_list import get_models_with_prefix, list_models
 from commands.model_create import create_model_folders
 
-def merge_models(target_model: str, source_models:list=[]):
+def merge_models(target_model: str, source_models:list=[], omit_test_data=False):
     """
     Merge one or more models training data into another one.
     :param target_model:str
@@ -22,13 +22,15 @@ def merge_models(target_model: str, source_models:list=[]):
             raise ValueError(f"Model '{source_model}' does not exist")
         for traindir_name in training_data_dirs:
             traindir_path = os.path.join(source_dir, traindir_name)
+            testdir_path = os.path.join(source_dir, "TEST_" + traindir_name)
             for curr_file in os.listdir(traindir_path):
                 if curr_file.startswith("."): continue
+                if omit_test_data and os.path.exists(os.path.join(testdir_path, curr_file)): continue
                 file_path = os.path.join(traindir_path, curr_file)
                 target_path = os.path.join(target_dir, traindir_name)
                 shutil.copy(file_path, target_path)
 
-def execute(target: str, models:list=[]):
+def execute(target: str, models:list=[], omit_test_data=False):
     """
     :param target:str
     :param mdls:list List of model names. If a model name ends with "*", it is used as
@@ -56,4 +58,4 @@ def execute(target: str, models:list=[]):
     if answer != "y":
         sys.exit(0)
 
-    merge_models(target, mdls)
+    merge_models(target, mdls, omit_test_data)
